assertHttpsUrl
