+++ src/utils/enrichAxiosErrors.ts
@@ -23,12 +23,12 @@ import {
   NO_INTERNET_MESSAGE,
   NO_RESPONSE_MESSAGE,
 } from "@/errors/errorHelpers";
-import { assertHttpsUrl } from "@/utils";
 import {
   ClientNetworkError,
   ClientNetworkPermissionError,
   RemoteServiceError,
 } from "@/errors/clientRequestErrors";
+import { assertHttpsUrl } from "@/errors/assertHttpsUrl";
 
 // eslint-disable-next-line prefer-destructuring -- It breaks EnvironmentPlugin
 const SERVICE_URL = process.env.SERVICE_URL;
